Move files into right location. No idea how this happened.
