--- __init__.py
@@ -1,105 +1,0 @@
-from __future__ import (
-    unicode_literals,
-    print_function,
-    absolute_import,
-    division,
-)
-
-from .pins import (
-    Factory,
-    Pin,
-    SPI,
-)
-from .pins.data import (
-    PiBoardInfo,
-    HeaderInfo,
-    PinInfo,
-    pi_info,
-)
-# Yes, import * is naughty, but exc imports nothing else so there's no cross
-# contamination here ... and besides, have you *seen* the list lately?!
-from .exc import *
-from .devices import (
-    Device,
-    GPIODevice,
-    CompositeDevice,
-)
-from .mixins import (
-    SharedMixin,
-    SourceMixin,
-    ValuesMixin,
-    EventsMixin,
-    HoldMixin,
-)
-from .input_devices import (
-    InputDevice,
-    DigitalInputDevice,
-    SmoothedInputDevice,
-    Button,
-    LineSensor,
-    MotionSensor,
-    LightSensor,
-    DistanceSensor,
-)
-from .spi_devices import (
-    SPIDevice,
-    AnalogInputDevice,
-    MCP3001,
-    MCP3002,
-    MCP3004,
-    MCP3008,
-    MCP3201,
-    MCP3202,
-    MCP3204,
-    MCP3208,
-    MCP3301,
-    MCP3302,
-    MCP3304,
-)
-from .output_devices import (
-    OutputDevice,
-    DigitalOutputDevice,
-    PWMOutputDevice,
-    PWMBuzzer,
-    PWMLED,
-    LED,
-    Buzzer,
-    Motor,
-    PhaseEnableMotor,
-    Servo,
-    AngularServo,
-    RGBLED,
-)
-from .boards import (
-    CompositeOutputDevice,
-    ButtonBoard,
-    LEDCollection,
-    LEDBoard,
-    LEDBarGraph,
-    LedBorg,
-    PiLiter,
-    PiLiterBarGraph,
-    TrafficLights,
-    PiTraffic,
-    PiStop,
-    StatusZero,
-    StatusBoard,
-    SnowPi,
-    TrafficLightsBuzzer,
-    FishDish,
-    TrafficHat,
-    Robot,
-    RyanteckRobot,
-    CamJamKitRobot,
-    PhaseEnableRobot,
-    PololuDRV8835Robot,
-    Energenie,
-    JamHat,
-)
-from .other_devices import (
-    InternalDevice,
-    PingServer,
-    CPUTemperature,
-    LoadAverage,
-    TimeOfDay,
-)
