+++ gpiozero/output_devices.py
@@ -503,6 +503,243 @@ class PWMOutputDevice(OutputDevice):
             self._write(value)
             if self._blink_thread.stopping.wait(delay):
                 break
+            
+class PWMBuzzer(PWMOutputDevice):
+    """
+    Extends :class: `PWMOutputDevice` and represents a piezo or tonal buzzer 
+    class.
+
+    :param: int pin:
+        The GPIO pin which the buzzer is attached to. See :ref:`pin-numbering` for
+        valid pin numbers.
+    
+    :param bool active_high:
+        If ``True`` (the default), the :meth:`on` method will set the GPIO to 
+        HIGH. If ``False``, the :meth:`on` method will set the GPIO to LOW (the
+        :meth:`off` method always does the opposite).
+    
+    :param float initial_value:
+        If ``0`` (the default), the buzzer will be off initially. Other values
+        between 0 and 1 can be specified as an initial value for the buzzer.
+        Note that ``None`` cannot be specified (unlike the parent class) as
+        there is no way to tell PWM not to alter the state of the pin.
+
+    :param float frequency:
+        The frequency (in Hz) of pulses emitted to drive the buzzer. Defaults
+        to 100Hz.
+    
+    :param Factory pin_factory:
+        See :doc:`api_pins` for more information (this is an advanced feature
+        which most users can ignore).
+    """
+
+    def __init__(
+            self, pin=None, active_high=True, initial_value=0, frequency=100,
+            pin_factory=None):
+        super(PWMBuzzer, self).__init__(
+            pin=pin, active_high=active_high, initial_value=initial_value, frequency=frequency,
+            pin_factory=pin_factory
+        )
+
+        self.notes = {
+            """
+            Dictionary containing musical notes which can be played through the PWMBuzzer.
+            """
+            'A0': 27.50,
+            'AS0': 29.14,
+            'B0': 30.87,
+            'C1': 32.70,
+            'CS1': 34.65,
+            'D1': 36.71,
+            'DS1': 38.89,
+            'E1': 41.20,
+            'F1': 43.65,
+            'FS1': 46.25,
+            'G1': 49.0,
+            'GS1': 51.91,
+            'A1': 55.00,
+            'AS1': 58.27,
+            'B1': 61.74,
+            'C2': 65.41,
+            'CS2': 69.30,
+            'D2': 73.42,
+            'DS2': 77.78,
+            'E2': 82.41,
+            'F2': 87.31,
+            'FS2': 92.50,
+            'G2': 98.0,
+            'GS2': 103.83,
+            'A2': 110.0,
+            'AS2': 116.54,
+            'B2': 123.47,
+            'C3': 130.81,
+            'CS3': 138.59,
+            'D3': 146.83,
+            'DS3': 155.56,
+            'E3': 164.81,
+            'F3': 174.61,
+            'FS3': 185.0,
+            'G3': 196.0,
+            'GS3': 207.65,
+            'A3': 220.0,
+            'AS3': 233.08,
+            'B3': 246.94,
+            'C4': 261.63,
+            'CS4': 277.18,
+            'D4': 293.66,
+            'DS4': 311.13,
+            'E4': 329.63,
+            'F4': 349.23,
+            'FS4': 369.99,
+            'G4': 392.0,
+            'GS4': 415.30,
+            'A4': 440.0,
+            'AS4': 466.16,
+            'B4': 493.88,
+            'C5': 523.25,
+            'CS5': 554.37,
+            'D5': 587.33,
+            'DS5': 622.25,
+            'E5': 659.26,
+            'F5': 698.46,
+            'FS5': 739.99,
+            'G5': 783.99,
+            'GS5': 830.61,
+            'A5': 880.00,
+            'AS5': 932.33,
+            'B5': 987.77,
+            'C6': 1046.50,
+            'CS6': 1108.73,
+            'D6': 1174.66,
+            'DS6': 1244.51,
+            'E6': 1318.51,
+            'F6': 1396.91,
+            'FS6': 1479.98,
+            'G6': 1567.98,
+            'GS6': 1661.22,
+            'A6': 1760.0,
+            'AS6': 1864.66,
+            'B6': 1975.53,
+            'C7': 2093.0,
+            'CS7': 2217.46,
+            'D7': 2349.32,
+            'DS7': 2489.02,
+            'E7': 2637.02,
+            'F7': 2793.82,
+            'FS7': 2959.96,
+            'G7': 3135.96,
+            'GS7': 3322.44,
+            'A7': 3520.0,
+            'AS7': 3729.31,
+            'B7': 3951.07,
+            'C8': 4186.01,
+        }
+
+        self.midi_notes = {
+            """
+            Dictionary containing MIDI notes which can be played through the PWMBuzzer.
+            """
+            '21': 27.50,
+            '22': 29.14,
+            '23': 30.87,
+            '24': 32.70,
+            '25': 34.65,
+            '26': 36.71,
+            '27': 38.89,
+            '28': 41.20,
+            '29': 43.65,
+            '30': 46.25,
+            '31': 49.0,
+            '32': 51.91,
+            '33': 55.00,
+            '34': 58.27,
+            '35': 61.74,
+            '36': 65.41,
+            '37': 69.30,
+            '38': 73.42,
+            '39': 77.78,
+            '40': 82.41,
+            '41': 87.31,
+            '42': 92.50,
+            '43': 98.0,
+            '44': 103.83,
+            '45': 110.0,
+            '46': 116.54,
+            '47': 123.47,
+            '48': 130.81,
+            '49': 138.59,
+            '50': 146.83,
+            '51': 155.56,
+            '52': 164.81,
+            '53': 174.61,
+            '54': 185.0,
+            '55': 196.0,
+            '56': 207.65,
+            '57': 220.0,
+            '58': 233.08,
+            '59': 246.94,
+            '60': 261.63,
+            '61': 277.18,
+            '62': 293.66,
+            '63': 311.13,
+            '64': 329.63,
+            '65': 349.23,
+            '66': 369.99,
+            '67': 392.0,
+            '68': 415.30,
+            '69': 440.0,
+            '70': 466.16,
+            '71': 493.88,
+            '72': 523.25,
+            '73': 554.37,
+            '74': 587.33,
+            '75': 622.25,
+            '76': 659.26,
+            '77': 698.46,
+            '78': 739.99,
+            '79': 783.99,
+            '80': 830.61,
+            '81': 880.0,
+            '82': 932.33,
+            '83': 987.77,
+            '84': 1046.50,
+            '85': 1108.73,
+            '86': 1174.66,
+            '87': 1244.51,
+            '88': 1318.51,
+            '89': 1396.91,
+            '90': 1479.98,
+            '91': 1567.98,
+            '92': 1661.22,
+            '93': 1760.0,
+            '94': 1864.66,
+            '95': 1975.53,
+            '96': 2093.0,
+            '97': 2217.46,
+            '98': 2349.32,
+            '99': 2489.02,
+            '100': 2637.02,
+            '101': 2793.82,
+            '102': 2959.96,
+            '103': 3135.96,
+            '104': 3322.44,
+            '105': 3520.0,
+            '106': 3729.31,
+            '107': 3951.07,
+            '108': 4186.01,
+        }
+
+        def play(self, note):
+            """
+            PWMBuzzer function to play a given note from the two available dictionaries.
+            """
+            if type(note) == int:
+                self.frequency = self.midi_notes[note]
+            else:
+                self.frequency = self.notes[note]
+            self.value = 0.5
+
+
 
 
 class PWMLED(PWMOutputDevice):
